Refactor commonalities between environment, solvegroup and groupedenvironment in a trait
