+++ src/cli/list.rs
@@ -12,6 +12,7 @@ use serde::Serialize;
 use uv_distribution::RegistryWheelIndex;
 
 use crate::lock_file::{UpdateLockFileOptions, UvResolutionContext};
+use crate::project::combine_feature::CombineFeature;
 use crate::pypi_tags::{get_pypi_tags, is_python_record};
 use crate::Project;
