+++ src/cli/tree.rs
@@ -7,6 +7,7 @@ use itertools::Itertools;
 use rattler_conda_types::Platform;
 
 use crate::lock_file::UpdateLockFileOptions;
+use crate::project::combine_feature::CombineFeature;
 use crate::Project;
 
 /// Show a tree of project dependencies
