+++ src/lock_file/update.rs
@@ -1,4 +1,5 @@
 use crate::lock_file::{PypiRecord, UvResolutionContext};
+use crate::project::combine_feature::CombineFeature;
 use crate::project::grouped_environment::GroupedEnvironmentName;
 use crate::pypi_mapping::{self, Reporter};
 use crate::pypi_marker_env::determine_marker_environment;
