+++ src/repodata.rs
@@ -1,4 +1,5 @@
 use crate::config::Config;
+use crate::project::combine_feature::CombineFeature;
 use crate::project::Environment;
 use crate::{config, progress, project::Project};
 use futures::{stream, StreamExt, TryStreamExt};
